Version 0.0.1 - Initial Upload
Code is very rough. Does not work. But the basics are there for me to start researching more into syntax and options. Not a seasoned Python programmer. Actually, this is my second "attempt at learning" using online instructions. If you are seasoned in programming in Python, you are free to laugh at me but I'm just learning / trying at this point.

This script is designed to shut down Microsoft Teams, clear the caches, and relaunch Teams. Unfortunately Teams doesn't have this feature built in so when things so weird, we need to do it manually at work.

This will become a quick script I can hopefully execute remotely or run on someone's machine to quickly do what I need to do.

Known Issues:
 - Teams executes with properties. All attempts to modify the shortcut have failed in trying to transfer knowledge from PHP over to Python. Left as is to look at later.
 - Could probably compress all the delete branches into the same TRY statement. Would be good to stop it if any end up failing. Will probably compress in next release